site/requirements.txt
git+https://github.com/blocher/python-scriptures
git+https://github.com/jazzband/django-robots
git+https://github.com/ivanchenkodmitry/django-meta@patch-1
git+https://github.com/blocher/django-inlinecss@django4
git+https://github.com/mailchimp/mailchimp-marketing-python.git
arrow==1.2.2
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.10.0
black==22.1.0
bugsnag==4.2.0
click==8.0.4
Delorean==1.0.0
Django==4.0.2
django-address==0.2.5
django-admin-sortable2==1.0.4
django-array-tags==0.2.2
django-cachekiller==0.7.1
django-ckeditor==6.2.0
django-cors-headers==3.11.0
django-debug-toolbar==3.2.4
django-distill==2.9.2
django-dotenv==1.4.2
django-environ==0.8.1
django-extensions==3.1.5
django-filter==21.1
django-js-asset==2.0.0
django-inlinecss
django-kronos==1.0
django-mathfilters==1.0.0
django-meta
django-mjml==0.12.0
django-model-utils==4.2.0
django-richtextfield==1.6.1
django-robots
django-sslserver==0.22
django-taggit==2.1.0
django_tinymce==3.4.0
django-webpack-loader==1.4.1
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
docx==0.2.4
drf-yasg==1.20.0
google-api-core==2.5.0
google-api-python-client==2.37.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.54.0
gunicorn==20.1.0
html2text==2020.1.16
humanize==4.0.0
Hypercorn==0.13.2
icalendar==4.0.9
indexed==1.2.1
ipython==8.0.1
kronos==0.2.2
lxml==4.8.0
Markdown==3.3.6
MarkupSafe==2.1.0
mjml==0.6.2
num2words==0.5.10
Pillow==9.0.1
pip==22.0.3
pipdeptree==2.2.1
pipenv==2022.1.8
platformdirs==2.5.1
pre-commit==2.17.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
pymemcache==3.5.0
pyrankvote==2.0.3
python-crontab==2.6.0
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==0.19.2
python-memcached==1.59
python-scriptures
pytz==2021.3
rake-nltk==1.0.6
requests==2.27.1
setuptools==60.9.3
tqdm==4.62.3
uvicorn==0.17.5
uvloop==0.16.0
Werkzeug==2.0.3
wheel==0.37.1
whitenoise==6.0.0
